backend/requirements.txt
# Core AI/ML Libraries
langchain==0.3.0
langchain-anthropic==0.3.5
langchain-groq==0.2.3
langchain-openai==0.3
langgraph==0.2.56

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
python-dotenv==1.0.0
colorama>=0.4.6
questionary>=2.1.0
rich>=13.9.4
tabulate>=0.9.0

# Visualization
matplotlib>=3.9.2

# External APIs
praw>=7.7.1
yfinance>=0.2.59
requests>=2.31.0
urllib3>=2.0.0
curl-cffi>=0.5.10

# Type hints and utilities
python-dateutil>=2.8.0
pydantic>=2.0.0
typing-extensions>=4.0.0 